Changined int to float

+++ Mike/C/data.h
@@ -1,4 +1,4 @@
-extern const int de00[];
-extern const int e01[];
-extern const int e10[];
+extern const float de00[];
+extern const float e01[];
+extern const float e10[];
 extern const int N;
